Workaround a ruby 2.3 issue with assigned constant names during autoload
The bug affects the constant that is being autoloaded and only
when it is assigned rather than declared with the `module` or `class`
keywords.

This can be reproduced with two files:

```ruby
autoload :Outer, './outer.rb'
puts Outer.name.inspect
```

```ruby
Outer = Class.new
puts Outer.name.inspect
```

which outputs this on ruby 2.3

```
nil
"Outer"
```

because the first Outer.name happens inside the autoload.

The workaround is to keep track of the full constant name in the
constant reference so we can use to find the unloaded namespaces.